Housekeeping: Move SVUnit helper classes into a Python package.
Makes test code cleaner and opens the door for future reuse.

- modified:   test/conftest.py
  - Move SVUnit class out.
  - Import it instead.

- new file:   test/resources/helpers/svunit/pyproject.toml
  - Config file enables us to `pip install -e` the `svunit` package.

- new file:   test/resources/helpers/svunit/src/svunit/__init__.py
  - Identifies `svunit` as a package.

- new file:   test/resources/helpers/svunit/src/svunit/svunit.py
  - SVUnit helper class source code.
